Moved API to services directory
