+++ src/popup/state/actions.ts
@@ -1,6 +1,6 @@
 import { AnimeStatus } from "../../model";
 import UserInfo from "../../model/userinfo";
-import { AnimeListEntry, SeriesUpdate } from "../../model/api";
+import { AnimeListEntry, SeriesUpdate } from "../../services/api";
 import AsyncDispatcher from "./asyncDispatcher";
 import SeriesInfo from "../../model/seriesinfo";
 import { ThemeData } from "../../model/theme";
